medik_cloud_deploy/requirements.txt
django==4.2.10
requests==2.31.0
debugpy # Required for debugging.

psycopg2-binary~=2.9.9
fpdf~=1.7.2
reportlab~=4.1.0
utils~=1.0.2

django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
